memtier: crc_verify SET:GET ratio
In the current implementation of the crc verify, the only SET:GET
ratio that is supported is 1:0. Thus, ratio configuration is
disallowed and the ratio is always set to 1:0.
This determines the request pattern when running the benchmark -
only SET requests will be sent.
In the data verification phase, a GET request would be generated
for each of the keys that were set and the returned value
will be verified.

Issue: LBM1-3307