memtier_benchmark: add compression-ratio parameter
This parameter will determine how much of the data should be
compressible (in case of random data).
For example, setting compression-ratio=50 will result in values, such
that for each value 50% of the data is random and 50% of the data is
compressible (i.e. constant value).
This will give us fine-grained control on the data, and allow us to
bench mark servers that support compression of the data.
Currently, this parameter is not in used. Its usage will be introduced
in later commits.

Issue: LBM1-3458