obj_gen: allow generating partially compressible data
Up until now, the object generator allowed generating random data or
constant data. But sometimes we would like to generate data that is part
random, part compressible, thus controlling how much of the data is
compressible.
In this commit we extend the object generator to support this feature.
We add a compression ratio member to the object generator. The object
generator will use it while allocating the value buffer - the size of
the random part (and the compressible part accordingly) will be
determined by the compression ratio.
Every time a new object is generated, the object generator will
calculate the required size of the random part and the compressible
part, and return a pointer to the place in the value buffer that respect
the ratio between the random part and the compressible part.

Issue: LBM1-3458