Client: if we dont use multi key get for crc verify we should use the request key.
When memtier issues a regular (not multi key get) get request ,
it actually sends a request with a GET opcode hence no key will be returned in the response.
since we use the key and the value to create the crc we will need the key when we verify the crc.
If we have only one value returned in the response and there is no key saved in the values list,
then we can assume this is a GET response and the key for the crc will be taken from
the request key list.
We have to check both the number of returned values and that the returned value dont have a key saved
with him as it is possible to set a multi key get requests with one key each,
which is a GETK request.

ISSUE: LBM1-3876