Add rate limiting to control the number of requests per second (#237)
* Add rate limiting to control the number of requests per second

A new 'rate-limiting' option was added to control the number of
request per second.

The rate limiting is based on the 'Token Bucket' algorithm, and
according to the configured rate, on each interval, a "new" amount
of requests allowed to be sent to the server.

The rate-limiting is at the connection level. Therefore, in cluster mode,
the limitation is for each shard, so if, for example, the cluster has three
shards and the user configured one request per second. On every second,
memtier-benchmark will send three requests, one for each shard.

* Added tests to cover --rate-limiting option. Ensured the help message explains the per connection rate-limit

* Fixed per cluster rps test limits

---------

Co-authored-by: YaacovHazan <yaacov.hazan@redis.com>
Co-authored-by: filipecosta90 <filipecosta.90@gmail.com>